+++ packages/cli/src/utils/compiler/compileBlock.ts
@@ -1,0 +1,48 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+import react from '@vitejs/plugin-react';
+import { build } from 'vite';
+import { viteExternalsPlugin } from 'vite-plugin-externals';
+
+import { getAppBridgeVersion } from '../appBridgeVersion';
+
+import { type CompilerOptions } from './compiler';
+
+export const compileBlock = async ({ projectPath, entryFile, outputName }: CompilerOptions) => {
+    const appBridgeVersion = getAppBridgeVersion(projectPath);
+    return build({
+        plugins: [
+            react(),
+            viteExternalsPlugin({
+                react: 'React',
+                'react-dom': 'ReactDOM',
+            }),
+        ],
+        define: {
+            'process.env.NODE_ENV': JSON.stringify('production'),
+        },
+        root: projectPath,
+        build: {
+            lib: {
+                name: outputName,
+                entry: entryFile,
+                formats: ['iife'],
+                fileName: () => 'index.js',
+            },
+            rollupOptions: {
+                external: ['react', 'react-dom'],
+                output: {
+                    globals: {
+                        react: 'React',
+                        'react-dom': 'ReactDOM',
+                    },
+                    footer: `
+                        window.${outputName} = ${outputName};
+                        window.${outputName}.dependencies = window.${outputName}.packages || {};
+                        window.${outputName}.dependencies['@frontify/app-bridge'] = '${appBridgeVersion}';
+                    `,
+                },
+            },
+        },
+    });
+};
