+++ packages/cli/src/utils/compiler/compiler.ts
@@ -1,0 +1,7 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export type CompilerOptions = {
+    projectPath: string;
+    entryFile: string;
+    outputName: string;
+};
