+++ packages/cli/src/utils/index.ts
@@ -1,7 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './appBridgeVersion';
-export * from './compiler';
+export * from './compiler/compiler';
 export * from './configuration';
 export * from './date';
 export * from './file';
@@ -15,3 +15,6 @@ export * from './reactiveJson';
 export * from './url';
 export * from './user';
 export * from './zip';
+export { compilePlatformApp } from './compiler/platformAppCompiler';
+export { compileBlock } from './compiler/compileBlock';
+export { compileTheme } from './compiler/compileTheme';
